Reset password email and premium role user added

+++ src/services/products.services.js
@@ -1,4 +1,5 @@
 import { productResponseDto } from "../dto/product-response.dto.js";
+import customErrors from "../errors/customErrors.js";
 import productsRepository from "../persistences/mongo/repositories/products.repository.js"
 
 const getAll = async (query, options) => {
@@ -12,17 +13,31 @@ const getById = async (id) => {
   return product;
 }
 
-const create = async (data) => {
-  const product = await productsRepository.create(data);
+const create = async (data, user) => {
+  let productData = data;
+  if(user.role === "premium"){
+    productData = {...data, owner: user._id};
+  };
+
+
+  const product = await productsRepository.create(productData);
   return product;
 }
 
-const update = async (id, data) => {
+const update = async (id, data, user) => {
+  const productData = await productsRepository.getById(id);
+  if(user.role === "premium" && productData.owner !== user._id){
+    throw customErrors.unauthorizedError("User not authorized");
+  }
   const product = await productsRepository.update(id, data);
   return product;
 }
 
-const deleteOne = async (id) => {
+const deleteOne = async (id, user) => {
+  const productData = await productsRepository.getById(id);
+  if(user.role === "premium" && productData.owner !== user._id){
+    throw customErrors.unauthorizedError("User not authorized");
+  }
   const product = await productsRepository.deleteOne(id);
   return product;
 }
